requirements.txt
certifi==2020.6.20
chardet==3.0.4
Django==2.1.4
djangorestframework==3.11.1
idna==2.10
mysqlclient==1.4.6
PyMySQL==0.9.3
pytz==2020.1
redis==3.5.3
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
